% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrack.R
\name{plotTrack}
\alias{plotTrack}
\title{plotTrack}
\usage{
plotTrack(
  gtfFile = NULL,
  gene = NULL,
  bigwigFile = NULL,
  multiple = FALSE,
  myTransId = NULL,
  strucCol = "#336699",
  arrowCol = "#336699",
  strucHeight = 0.25,
  uped = 1000,
  downed = 1000,
  addfacetCol = FALSE,
  facetFill = NULL,
  borderCol = NULL,
  sampleAes = NULL,
  trackCol = NULL,
  facetVars = NULL,
  relHeight = 0.1,
  annoTextSize = 10,
  base_size = 12,
  ylab = "Read coverage"
)
}
\arguments{
\item{gtfFile}{Your annotation file with GTF format, eg 'Mus_musculus.GRCm38.102.gtf'.}

\item{gene}{Which gene used to plot, eg 'Actb'.}

\item{bigwigFile}{loadBWfile function output results.}

\item{multiple}{Multiple gene isoforms will be ploted when it is TRUE, default is 'FALSE'.}

\item{myTransId}{Which transcript will be plot, can be multiple vectors. Parameter 'multiple' should be TRUE when you want to plot 2 or more transcripts.}

\item{strucCol}{Gene colors, defult is '#336699'.}

\item{arrowCol}{Arrow colors on gene structure, defult is '#336699'.}

\item{strucHeight}{Gene structure height, defult is 0.25.}

\item{uped}{Extend from start site, defult is 1000 bp.}

\item{downed}{Extend from end site, defult is 1000 bp.}

\item{addfacetCol}{Whether add more facets, defult is 'FALSE'.}

\item{facetFill}{Facets fill colors.}

\item{borderCol}{Facets border colors.}

\item{sampleAes}{Which variable should be used to aes track plot.}

\item{trackCol}{Supply your own colors vectors to change track colors.}

\item{facetVars}{Facet by this variable names. Can be multiple.}

\item{relHeight}{Relative height of the gene structure to the track plot, default is 0.1.}

\item{annoTextSize}{Facet text size, default is 10.}

\item{base_size}{ggplot base size, default is 12.}

\item{ylab}{Y axis label, default is 'Read coverage'.}
}
\value{

}
\description{
plotTrack is used to visulaize gene track from bigwig files.
}
\examples{
\dontrun{
  # examples
  plotTrack(gtfFile = gtf,
            gene = 'Actb',
            arrowCol = 'black',
            bigwigFile = allBw,
            sampleAes = 'sample',
            facetVars = 'name')

  # change aes variable
  plotTrack(gtfFile = gtf,
            gene = 'Actb',
            arrowCol = 'black',
            bigwigFile = allBw,
            sampleAes = 'name',
            facetVars = 'name')

  # change track colors
  plotTrack(gtfFile = gtf,
            gene = 'Actb',
            arrowCol = 'black',
            bigwigFile = allBw,
            sampleAes = 'name',
            facetVars = 'name',
            trackCol = pal_lancet()(6))

  # draw multiple genes
  plotTrack(gtfFile = gtf,
            gene = 'Tnf',
            arrowCol = 'black',
            bigwigFile = allBw,
            sampleAes = 'sample',
            facetVars = 'name',
            multiple = TRUE,
            myTransId = c('ENSMUST00000025263','ENSMUST00000167924'))

  # change facet fill colors
  plotTrack(gtfFile = gtf,
            gene = 'Actb',
            arrowCol = 'black',
            bigwigFile = allBw,
            sampleAes = 'sample',
            addfacetCol = TRUE,
            facetVars = 'name',
            facetFill = pal_d3()(6),
            borderCol = rep('white',6))

  # add one facet
  plotTrack(gtfFile = gtf,
            gene = 'Actb',
            arrowCol = 'black',
            bigwigFile = allBw,
            sampleAes = 'sample',
            facetVars = c('group','sample'),
            addfacetCol = TRUE,
            facetFill = c(pal_d3()(3),pal_npg()(6)),
            borderCol = rep('white',9))

  # add more one facet
  plotTrack(gtfFile = gtf,
            gene = 'Actb',
            arrowCol = 'black',
            bigwigFile = allBw,
            sampleAes = 'sample',
            facetVars = c('group','sample','name'),
            addfacetCol = TRUE,
            facetFill = c(pal_d3()(3),pal_npg()(6),pal_locuszoom()(6)),
            borderCol = rep('white',15))
}
}
